requirements.txt
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scikit-surprise==1.1.3
mlflow==2.9.2
apache-airflow==2.8.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
PyYAML==6.0.1
click==8.1.7
pytest==7.4.3
pytest-cov==4.1.0
sqlalchemy==2.0.23
